requirements.txt
py-cord==2.3.0
python-dotenv==0.21.0
pandas==1.5.2
rswiki-wrapper=0.0.2
matplotlib==3.6.2